+++ CHANGELOG.md
@@ -7,13 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet
+### Fixed
+
+- stpp box handles optional empty lists properly (a single zero byte)
 
 ## [0.34.1] - 2023-03-09
 
-### [Fixed]
+### Fixed
 
 - Only start new segment at start or styp box
+
+### Changed
+
+-
 
 ## [0.34.0] - 2023-02-28